chore: rename linebroadening to apodization
