+++ src/data/Filters.ts
@@ -1,4 +1,5 @@
 import * as absolute from './data1d/filter1d/absolute';
+import * as apodization from './data1d/filter1d/apodization';
 import * as autoPhaseCorrection from './data1d/filter1d/autoPhaseCorrection';
 import * as baselineCorrection from './data1d/filter1d/baselineCorrection';
 import * as centerMean from './data1d/filter1d/centerMean';
@@ -7,7 +8,6 @@ import * as equallySpaced from './data1d/filter1d/equallySpaced';
 import * as exclusionZones from './data1d/filter1d/exclusionZones';
 import * as fft from './data1d/filter1d/fft';
 import * as fromTo from './data1d/filter1d/fromTo';
-import * as lineBroadening from './data1d/filter1d/lineBroadening';
 import * as pareto from './data1d/filter1d/pareto';
 import * as phaseCorrection from './data1d/filter1d/phaseCorrection';
 import * as shiftX from './data1d/filter1d/shiftX';
@@ -20,7 +20,7 @@ import * as shift2DY from './data2d/filter2d/shiftY';
 export {
   absolute,
   baselineCorrection,
-  lineBroadening,
+  apodization,
   zeroFillingAndApodization,
   fft,
   phaseCorrection,
