+++ src/data/data1d/filter1d/__tests__/apodization.test.ts
@@ -1,28 +1,29 @@
-import { apply } from '../lineBroadening';
+import { apply } from '../apodization';
 
 describe('lineBroadening', () => {
   it('simple x, re, im to 1 Hz exp.', () => {
     let spectrum = {
       data: {
-        re: new Float64Array([1, 1, 1, 1]),
+        re: new Float64Array([1, 2, 3, 4]),
         im: new Float64Array([0, 0, 0, 0]),
-        x: new Float64Array([0, 1, 2, 3]),
+        x: new Float64Array([10, 20, 30, 40]),
       },
       info: {
         isComplex: true,
         isFid: true,
       },
     };
-    apply(spectrum, 1);
+    apply(spectrum, {
+      lineBroadeningValue: 0.1,
+      gaussBroadeningValue: 0,
+      centerValue: 0,
+    });
     expect(spectrum.data).toStrictEqual({
       re: new Float64Array([
-        1,
-        0.06598803584531254, // this is ugly... may not work. toBeCloseTo not for objects
-        0.004354420874722253,
-        0.00028733968076674924,
+        1, 0.08642783652754452, 0.0056023281951239675, 0.0003227980702812185,
       ]),
       im: new Float64Array([0, 0, 0, 0]),
-      x: new Float64Array([0, 1, 2, 3]),
+      x: new Float64Array([10, 20, 30, 40]),
     });
   });
 });
