+++ src/data/data1d/filter1d/apodization.ts
